Change Slack MCP to use new pattern
